# perf(core): avoid persisting install state if it didn't change