**What's the problem this PR addresses?**

During an install Yarn serializes and compresses the install state twice when it hasn't changed, once before checking build scripts and once after everything is done, which with a large Project is expensive.

**How did you fix it?**

- Don't call `Project.persistInstallStateFile` unless a build actually needs to run
- Check if the serialized state is different from the previous one before compressing it

**Result**

Running `yarn install` in the Storybook repo (using PnP) when nothing has changed

```diff
- Done with warnings in 4s 748ms
+ Done with warnings in 3s 569ms
```

Gatsby benchmark
```diff
 $ bash scripts/bench-run.sh yarn gatsby $(mktemp -d)
 Benchmark #1: yarn add dummy-pkg@link:./dummy-pkg
-  Time (mean ± σ):      2.901 s ±  0.061 s    [User: 2.639 s, System: 0.523 s]
-  Range (min … max):    2.827 s …  2.985 s    10 runs
+  Time (mean ± σ):      2.751 s ±  0.061 s    [User: 2.544 s, System: 0.474 s]
+  Range (min … max):    2.657 s …  2.796 s    10 runs
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.